+++ src/utils/toolbox.ts
@@ -1,13 +1,13 @@
 export const random = {
 	float(max = 1, min = 0): number {
+		[min, max] = [Math.ceil(min), Math.floor(max)];
 		return Math.random() * (max - min) + min;
 	},
 	int(max = 1, min = 0): number {
-		[min, max] = [Math.ceil(min), Math.floor(max)];
 		return Math.floor(this.float(max, min));
 	},
 	bool(): boolean {
-		return !!this.int();
+		return this.float() < 0.5;
 	},
 	array(length: number, max: number, min = 0): Array<number> {
 		return Array.from({ length }, () => this.int(max, min));